fix(ui): reliable Windows launcher + UI flag; add start_windows_ui.bat, start_ui_only.bat, debug_start.bat
